requirements.txt
SQLAlchemy==0.8.2
bigfloat==0.3.0a2
blist==1.3.4
cdecimal==2.3
ecdsa==0.8
gmpy2==2.0.2
isodate==0.4.8
pycrypto==2.6
python-bitcoin==0.0.5
python-gflags==2.0
recordtype==1.1
requests==0.14.0
simplejson==2.6.2
six==1.3.0
